\name{gse2189}
\docType{data}
\alias{gse2189}
\title{
  Sample dataset for 'pyn' package.
}
\description{
  Sample dataset for the 'pyn' package: three replicate arrays of one of six samples from
  GSE2189, one of which contains significant spatial aberrations.
}
\usage{gse2189}
%- maybe also 'usage' for other objects documented here.
\format{
  An AffyBatch object, as loaded by affy::ReadAffy.
}
\source{
  Magda, D., et al. (2005). Motexafin gadolinium disrupts zinc metabolism in human cancer
  cell lines. Cancer Res, 65(9):3837-45. 
}
\keyword{ ~datasets }
